Codechange: Refactor dirty block system used by gfx engine.
The purpose of this refactoring is to increase the granularity with
which dirty blocks are tracked, while simultaneously improving the
documentation of the algorithm that is used.

This should decrease the size of the areas that are unneccessarily
redrawn, while still allowing small locallised changes to be
coalesced for efficient redrawing.